[REFACTOR] Separate functions
